Fitting (generalized) linear models to synthetic data

Fits generalized linear models or simple linear models to the synthesised data set(s) using glm and lm function respectively.

Usage

glm.synds(formula, family = "binomial", data,  ...)
lm.synds(formula, data, ...)

# S3 method for fit.synds
print(x, msel = NULL, ...)

Arguments

formula

a symbolic description of the model to be estimated. A typical model has the form response ~ predictors. See the documentation of glm and formula for details.

family

a description of the error distribution and link function to be used in the model. See the documentation of glm and family for details.

data

an object of class synds, which stands for 'synthesised data set'. It is typically created by function syn and it includes data$m synthesised data set(s).

...

additional parameters passed to glm or lm.

x

an object of class fit.synds.

msel

index or indices of synthetic data copies for which coefficient estimates are to be displayed. If NULL (default) the combined (average) coefficient estimates are printed.

Value

The summary function (summary.fit.synds) can be used to obtain the combined results of models fitted to each of the m

synthetic data sets.

An object of class fit.synds. It is a list with the following components:

call

the original call to glm.synds or lm.synds.

mcoefavg

combined (average) coefficient estimates.

mvaravg

combined (average) variance estimates of mcoef.

analyses

summary.glm or summary.lm object respectively or a list of m such objects.

fitting.function

function used to fit the model.

n

a number of cases in the original data.

k

a number of cases in the synthesised data.

proper

a logical value indicating whether synthetic data were generated using proper synthesis.

m

the number of synthetic versions of the observed data.

method

a vector of synthesising methods applied to each variable in the saved synthesised data.

incomplete

a logical value indicating whether the dependent variable in the model was not synthesised.

mcoef

a matrix of coefficients estimates from all m syntheses.

mvar

a matrix of variance estimates from all m syntheses.

See also

Examples

### Logit model 
ods <- SD2011[1:1000, c("sex", "age", "edu", "marital", "ls", "smoke")]
s1 <- syn(ods, m = 3)
#> 
#> Synthesis number 1
#> --------------------
#>  sex age edu marital ls smoke
#> 
#> Synthesis number 2
#> --------------------
#>  sex age edu marital ls smoke
#> 
#> Synthesis number 3
#> --------------------
#>  sex age edu marital ls smoke
f1 <- glm.synds(smoke ~ sex + age + edu + marital + ls, data = s1, family = "binomial")
f1
#> Note: To get more details of the fit see vignette on inference.
#> 
#> Call:
#> glm.synds(formula = smoke ~ sex + age + edu + marital + ls, family = "binomial", 
#>     data = s1)
#> 
#> Average coefficient estimates from 3 syntheses:
#>                 (Intercept)                   sexFEMALE 
#>                  0.80287558                  0.53496969 
#>                         age       eduVOCATIONAL/GRAMMAR 
#>                  0.02803355                 -0.15490561 
#>                eduSECONDARY eduPOST-SECONDARY OR HIGHER 
#>                  0.58332737                  0.63332274 
#>              maritalMARRIED              maritalWIDOWED 
#>                 -0.68032578                 -0.82307362 
#>             maritalDIVORCED    maritalLEGALLY SEPARATED 
#>                 -1.50435329                 -1.25034501 
#>   maritalDE FACTO SEPARATED                   lsPLEASED 
#>                 -6.45548853                 -0.72920642 
#>          lsMOSTLY SATISFIED                     lsMIXED 
#>                 -0.94639096                 -0.92814719 
#>       lsMOSTLY DISSATISFIED                   lsUNHAPPY 
#>                 -1.39566956                  2.86009047 
#>                  lsTERRIBLE 
#>                 -1.65803580 
print(f1, msel = 1:2)
#> Note: To get more details of the fit see vignette on inference.
#> 
#> Call:
#> glm.synds(formula = smoke ~ sex + age + edu + marital + ls, family = "binomial", 
#>     data = s1)
#> 
#> Coefficient estimates for selected synthetic data set(s):
#>       (Intercept) sexFEMALE        age eduVOCATIONAL/GRAMMAR eduSECONDARY
#> syn=1   0.4241985 0.6087996 0.02549054           -0.43325912    0.2352771
#> syn=2   2.1211847 0.3219266 0.03492592           -0.04602541    0.7159938
#>       eduPOST-SECONDARY OR HIGHER maritalMARRIED maritalWIDOWED maritalDIVORCED
#> syn=1                  0.01254626     -0.8543933     -1.0532286      -0.9747471
#> syn=2                  0.91947874     -0.9107269     -0.7128121      -1.8523790
#>       maritalLEGALLY SEPARATED maritalDE FACTO SEPARATED lsPLEASED
#> syn=1                -0.886029               -16.0269919  0.220863
#> syn=2                -1.614661                -0.9391807 -2.226010
#>       lsMOSTLY SATISFIED    lsMIXED lsMOSTLY DISSATISFIED  lsUNHAPPY lsTERRIBLE
#> syn=1         -0.1837651 -0.3352089            -0.7183869 -0.9336909  -1.100587
#> syn=2         -2.4054188 -2.3039977            -2.8978997 10.9227435  -1.608371
  
### Linear model
ods <- SD2011[1:1000,c("sex", "age", "income", "marital", "depress")]
ods$income[ods$income == -8] <- NA
s2 <- syn(ods, m = 3)
#> 
#> Synthesis number 1
#> --------------------
#>  sex age income marital depress
#> 
#> Synthesis number 2
#> --------------------
#>  sex age income marital depress
#> 
#> Synthesis number 3
#> --------------------
#>  sex age income marital depress
f2 <- lm.synds(depress ~ sex + age + log(income) + marital, data = s2)
f2
#> Note: To get more details of the fit see vignette on inference.
#> 
#> Call:
#> lm.synds(formula = depress ~ sex + age + log(income) + marital, 
#>     data = s2)
#> 
#> Average coefficient estimates from 3 syntheses:
#>               (Intercept)                 sexFEMALE                       age 
#>                 6.4972123                 0.6091896                 0.1314323 
#>               log(income)            maritalMARRIED            maritalWIDOWED 
#>                -1.1586932                -0.4233951                 0.9264299 
#>           maritalDIVORCED maritalDE FACTO SEPARATED  maritalLEGALLY SEPARATED 
#>                 0.2871722                 0.4846360                 5.1089478 
print(f2,1:3)
#> Note: To get more details of the fit see vignette on inference.
#> 
#> Call:
#> lm.synds(formula = depress ~ sex + age + log(income) + marital, 
#>     data = s2)
#> 
#> Coefficient estimates for selected synthetic data set(s):
#>       (Intercept) sexFEMALE       age log(income) maritalMARRIED maritalWIDOWED
#> syn=1    6.618506 0.4903958 0.1367388   -1.216413     -0.2632835      1.1988759
#> syn=2    7.610038 0.3731267 0.1236601   -1.237654     -0.4406671      0.8884281
#> syn=3    5.263093 0.9640464 0.1338981   -1.022013     -0.5662346      0.6919859
#>       maritalDIVORCED maritalDE FACTO SEPARATED maritalLEGALLY SEPARATED
#> syn=1     -0.08829972                 0.9581832                       NA
#> syn=2      1.49427596                 0.8467576                 5.108948
#> syn=3     -0.54445964                -0.3510329                       NA