Fitting ordered logistic models to synthetic data

Fits ordered logistic models to the synthesised data set(s) using the polr function.

Usage

polr.synds(formula, data, ...)

Arguments

formula

a symbolic description of the model to be estimated. A typical model has the form response ~ predictors. See the documentation of polr and formula for details.

data

an object of class synds, which stands for 'synthesised data set'. It is typically created by function syn and it includes data$m synthesised data set(s).

...

additional parameters passed to polr.

Value

To print the results the print function (print.fit.synds) can be used. The summary function (summary.fit.synds) can be used to obtain the combined results of models fitted to each of the

m synthetic data sets.

An object of class fit.synds. It is a list with the following components:

call

the original call to polr.synds.

mcoefavg

combined (average) coefficient estimates.

mvaravg

combined (average) variance estimates of mcoef.

analyses

an object summarising the fit to each synthetic data set or a list of m such objects. Note that this is different from the object created by summary.polr to make it compatible with other fitting methods for the compare.fit.synds and summary.fit.synds functions. In particular the coefficients combine the items coefficients and zeta from summary.polr to make a combined vector of coefficients.

fitting.function

function used to fit the model.

n

a number of cases in the original data.

k

a number of cases in the synthesised data.

proper

a logical value indicating whether synthetic data were generated using proper synthesis.

m

the number of synthetic versions of the observed data.

method

a vector of synthesising methods applied to each variable in the saved synthesised data.

incomplete

a logical value indicating whether the dependent variable in the model was not synthesised.

mcoef

a matrix of coefficients estimates from all m syntheses.

mvar

a matrix of variance estimates from all m syntheses.

See also

Examples

ods <- SD2011[1:1000, c("sex", "age", "edu", "marital", "ls", "smoke")]
s1 <- syn(ods, m = 3)
#> 
#> Synthesis number 1
#> --------------------
#>  sex age edu marital ls smoke
#> 
#> Synthesis number 2
#> --------------------
#>  sex age edu marital ls smoke
#> 
#> Synthesis number 3
#> --------------------
#>  sex age edu marital ls smoke
f1 <- polr.synds(edu  ~ sex + age, data = s1)
summary(f1)
#> Fit to synthetic data set with 3 syntheses. Inference to coefficients
#> and standard errors that would be obtained from the original data.
#> 
#> Call:
#> polr.synds(formula = edu ~ sex + age, data = s1)
#> 
#> Combined estimates:
#>                                         xpct(Beta) xpct(se.Beta)  xpct(z)
#> sexFEMALE                                0.2994688     0.1162397   2.5763
#> age                                     -0.0254499     0.0033238  -7.6568
#> PRIMARY/NO EDUCATION|VOCATIONAL/GRAMMAR -2.6547455     0.1871712 -14.1835
#> VOCATIONAL/GRAMMAR|SECONDARY            -0.9262975     0.1659100  -5.5831
#> SECONDARY|POST-SECONDARY OR HIGHER       0.5574158     0.1701544   3.2759
#>                                         Pr(>|xpct(z)|)    
#> sexFEMALE                                     0.009986 ** 
#> age                                          1.907e-14 ***
#> PRIMARY/NO EDUCATION|VOCATIONAL/GRAMMAR      < 2.2e-16 ***
#> VOCATIONAL/GRAMMAR|SECONDARY                 2.362e-08 ***
#> SECONDARY|POST-SECONDARY OR HIGHER            0.001053 ** 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
print(f1, msel = 1:2)
#> Note: To get more details of the fit see vignette on inference.
#> 
#> Call:
#> polr.synds(formula = edu ~ sex + age, data = s1)
#> 
#> Coefficient estimates for selected synthetic data set(s):
#>       sexFEMALE         age PRIMARY/NO EDUCATION|VOCATIONAL/GRAMMAR
#> syn=1 0.2347645 -0.02950875                               -2.886276
#> syn=2 0.3165962 -0.02356529                               -2.486376
#>       VOCATIONAL/GRAMMAR|SECONDARY SECONDARY|POST-SECONDARY OR HIGHER
#> syn=1                   -1.1112404                          0.3863834
#> syn=2                   -0.7542516                          0.6621760
compare(f1, ods)
#> Error in eval(expr, p): object 'ods' not found