Replications in synthetic data

Determines which unique units in the synthesised data set(s) replicates unique units in the original observed data set.

Usage

replicated.uniques(object, data, exclude = NULL)

Arguments

object

an object of class synds, which stands for 'synthesised data set'. It is typically created by function syn() and it includes object$m synthesised data set(s).

data

the original observed data set.

exclude

a single string or a vector of strings with name(s) of variable(s) to be excluded from the identification of uniques.

Value

A list with the following components:

replications

a vector (for object$m = 1) or a data frame with object$m columns (for object$m > 1) with logical values indicateing duplicates in mth synthetic data set.

no.replications

a single number or a vector of object$m integers indicating the number of duplicates in the synthetic data set(s).

no.uniques

a number of unique individuals in the original data set.

per.replications

a single number or a vector of object$m numeric values indicating the percentage of duplicates in the synthetic data set(s).

See also

Examples

ods <- SD2011[1:1000,c("sex","age","edu","marital","smoke")]
s1 <- syn(ods, m = 2)
#> 
#> Synthesis number 1
#> --------------------
#>  sex age edu marital smoke
#> 
#> Synthesis number 2
#> --------------------
#>  sex age edu marital smoke
replicated.uniques(s1,ods)
#> $replications
#>          [,1]  [,2]
#>    [1,]  TRUE FALSE
#>    [2,] FALSE FALSE
#>    [3,] FALSE FALSE
#>    [4,]  TRUE FALSE
#>    [5,]  TRUE FALSE
#>    [6,] FALSE FALSE
#>    [7,] FALSE FALSE
#>    [8,] FALSE FALSE
#>    [9,]  TRUE FALSE
#>   [10,] FALSE FALSE
#>   [11,] FALSE FALSE
#>   [12,] FALSE FALSE
#>   [13,] FALSE FALSE
#>   [14,] FALSE FALSE
#>   [15,] FALSE FALSE
#>   [16,] FALSE FALSE
#>   [17,] FALSE  TRUE
#>   [18,] FALSE FALSE
#>   [19,] FALSE FALSE
#>   [20,] FALSE FALSE
#>   [21,] FALSE FALSE
#>   [22,]  TRUE  TRUE
#>   [23,] FALSE FALSE
#>   [24,] FALSE  TRUE
#>   [25,] FALSE FALSE
#>   [26,]  TRUE FALSE
#>   [27,] FALSE FALSE
#>   [28,]  TRUE  TRUE
#>   [29,] FALSE FALSE
#>   [30,] FALSE  TRUE
#>   [31,] FALSE FALSE
#>   [32,] FALSE FALSE
#>   [33,] FALSE FALSE
#>   [34,]  TRUE FALSE
#>   [35,] FALSE FALSE
#>   [36,] FALSE FALSE
#>   [37,] FALSE FALSE
#>   [38,]  TRUE FALSE
#>   [39,] FALSE FALSE
#>   [40,] FALSE FALSE
#>   [41,] FALSE FALSE
#>   [42,] FALSE FALSE
#>   [43,] FALSE FALSE
#>   [44,] FALSE FALSE
#>   [45,] FALSE FALSE
#>   [46,] FALSE FALSE
#>   [47,] FALSE FALSE
#>   [48,]  TRUE FALSE
#>   [49,] FALSE FALSE
#>   [50,] FALSE FALSE
#>   [51,] FALSE FALSE
#>   [52,] FALSE FALSE
#>   [53,] FALSE FALSE
#>   [54,] FALSE FALSE
#>   [55,] FALSE FALSE
#>   [56,] FALSE FALSE
#>   [57,] FALSE FALSE
#>   [58,] FALSE  TRUE
#>   [59,] FALSE FALSE
#>   [60,]  TRUE FALSE
#>   [61,] FALSE FALSE
#>   [62,] FALSE FALSE
#>   [63,] FALSE FALSE
#>   [64,]  TRUE FALSE
#>   [65,] FALSE FALSE
#>   [66,] FALSE FALSE
#>   [67,] FALSE FALSE
#>   [68,] FALSE FALSE
#>   [69,] FALSE FALSE
#>   [70,] FALSE FALSE
#>   [71,]  TRUE FALSE
#>   [72,] FALSE FALSE
#>   [73,] FALSE FALSE
#>   [74,] FALSE FALSE
#>   [75,] FALSE FALSE
#>   [76,] FALSE FALSE
#>   [77,]  TRUE FALSE
#>   [78,] FALSE FALSE
#>   [79,] FALSE FALSE
#>   [80,] FALSE FALSE
#>   [81,] FALSE FALSE
#>   [82,] FALSE FALSE
#>   [83,] FALSE FALSE
#>   [84,] FALSE FALSE
#>   [85,] FALSE  TRUE
#>   [86,]  TRUE FALSE
#>   [87,] FALSE FALSE
#>   [88,] FALSE FALSE
#>   [89,] FALSE FALSE
#>   [90,] FALSE  TRUE
#>   [91,] FALSE FALSE
#>   [92,]  TRUE FALSE
#>   [93,] FALSE FALSE
#>   [94,] FALSE FALSE
#>   [95,] FALSE FALSE
#>   [96,] FALSE FALSE
#>   [97,] FALSE FALSE
#>   [98,] FALSE FALSE
#>   [99,] FALSE  TRUE
#>  [100,] FALSE FALSE
#>  [101,] FALSE FALSE
#>  [102,] FALSE FALSE
#>  [103,] FALSE FALSE
#>  [104,] FALSE  TRUE
#>  [105,] FALSE FALSE
#>  [106,] FALSE FALSE
#>  [107,] FALSE FALSE
#>  [108,]  TRUE FALSE
#>  [109,] FALSE FALSE
#>  [110,] FALSE FALSE
#>  [111,] FALSE FALSE
#>  [112,] FALSE FALSE
#>  [113,]  TRUE  TRUE
#>  [114,] FALSE FALSE
#>  [115,] FALSE  TRUE
#>  [116,] FALSE FALSE
#>  [117,] FALSE FALSE
#>  [118,] FALSE FALSE
#>  [119,] FALSE FALSE
#>  [120,] FALSE FALSE
#>  [121,]  TRUE FALSE
#>  [122,] FALSE FALSE
#>  [123,] FALSE FALSE
#>  [124,] FALSE FALSE
#>  [125,] FALSE FALSE
#>  [126,] FALSE FALSE
#>  [127,] FALSE FALSE
#>  [128,] FALSE FALSE
#>  [129,] FALSE FALSE
#>  [130,] FALSE FALSE
#>  [131,] FALSE FALSE
#>  [132,] FALSE FALSE
#>  [133,] FALSE FALSE
#>  [134,] FALSE  TRUE
#>  [135,] FALSE FALSE
#>  [136,] FALSE FALSE
#>  [137,] FALSE FALSE
#>  [138,] FALSE FALSE
#>  [139,]  TRUE FALSE
#>  [140,] FALSE FALSE
#>  [141,] FALSE FALSE
#>  [142,] FALSE FALSE
#>  [143,] FALSE FALSE
#>  [144,] FALSE FALSE
#>  [145,] FALSE FALSE
#>  [146,] FALSE FALSE
#>  [147,] FALSE FALSE
#>  [148,] FALSE FALSE
#>  [149,] FALSE FALSE
#>  [150,] FALSE FALSE
#>  [151,] FALSE FALSE
#>  [152,] FALSE FALSE
#>  [153,] FALSE FALSE
#>  [154,] FALSE FALSE
#>  [155,] FALSE FALSE
#>  [156,] FALSE FALSE
#>  [157,]  TRUE FALSE
#>  [158,] FALSE FALSE
#>  [159,] FALSE FALSE
#>  [160,]  TRUE FALSE
#>  [161,] FALSE FALSE
#>  [162,] FALSE FALSE
#>  [163,] FALSE FALSE
#>  [164,] FALSE FALSE
#>  [165,] FALSE FALSE
#>  [166,] FALSE FALSE
#>  [167,] FALSE FALSE
#>  [168,]  TRUE FALSE
#>  [169,] FALSE FALSE
#>  [170,] FALSE FALSE
#>  [171,] FALSE FALSE
#>  [172,] FALSE FALSE
#>  [173,] FALSE FALSE
#>  [174,] FALSE  TRUE
#>  [175,] FALSE FALSE
#>  [176,] FALSE FALSE
#>  [177,]  TRUE FALSE
#>  [178,] FALSE FALSE
#>  [179,] FALSE FALSE
#>  [180,] FALSE FALSE
#>  [181,] FALSE FALSE
#>  [182,] FALSE FALSE
#>  [183,] FALSE FALSE
#>  [184,] FALSE  TRUE
#>  [185,] FALSE FALSE
#>  [186,] FALSE FALSE
#>  [187,] FALSE FALSE
#>  [188,] FALSE FALSE
#>  [189,] FALSE FALSE
#>  [190,]  TRUE FALSE
#>  [191,] FALSE FALSE
#>  [192,] FALSE FALSE
#>  [193,] FALSE FALSE
#>  [194,] FALSE FALSE
#>  [195,] FALSE FALSE
#>  [196,] FALSE FALSE
#>  [197,] FALSE FALSE
#>  [198,] FALSE FALSE
#>  [199,] FALSE FALSE
#>  [200,] FALSE FALSE
#>  [201,] FALSE FALSE
#>  [202,] FALSE FALSE
#>  [203,] FALSE FALSE
#>  [204,] FALSE FALSE
#>  [205,] FALSE FALSE
#>  [206,] FALSE FALSE
#>  [207,] FALSE FALSE
#>  [208,] FALSE FALSE
#>  [209,] FALSE FALSE
#>  [210,] FALSE FALSE
#>  [211,] FALSE FALSE
#>  [212,] FALSE FALSE
#>  [213,] FALSE FALSE
#>  [214,] FALSE FALSE
#>  [215,] FALSE FALSE
#>  [216,] FALSE FALSE
#>  [217,] FALSE  TRUE
#>  [218,] FALSE FALSE
#>  [219,] FALSE FALSE
#>  [220,] FALSE FALSE
#>  [221,] FALSE FALSE
#>  [222,] FALSE FALSE
#>  [223,]  TRUE FALSE
#>  [224,] FALSE  TRUE
#>  [225,] FALSE FALSE
#>  [226,] FALSE FALSE
#>  [227,] FALSE FALSE
#>  [228,] FALSE FALSE
#>  [229,] FALSE  TRUE
#>  [230,] FALSE FALSE
#>  [231,] FALSE FALSE
#>  [232,] FALSE FALSE
#>  [233,] FALSE FALSE
#>  [234,]  TRUE FALSE
#>  [235,] FALSE FALSE
#>  [236,] FALSE FALSE
#>  [237,] FALSE FALSE
#>  [238,] FALSE FALSE
#>  [239,] FALSE FALSE
#>  [240,] FALSE FALSE
#>  [241,] FALSE FALSE
#>  [242,]  TRUE FALSE
#>  [243,] FALSE FALSE
#>  [244,] FALSE FALSE
#>  [245,] FALSE FALSE
#>  [246,] FALSE FALSE
#>  [247,] FALSE FALSE
#>  [248,] FALSE FALSE
#>  [249,] FALSE FALSE
#>  [250,] FALSE  TRUE
#>  [251,] FALSE FALSE
#>  [252,] FALSE  TRUE
#>  [253,] FALSE FALSE
#>  [254,] FALSE FALSE
#>  [255,] FALSE FALSE
#>  [256,] FALSE FALSE
#>  [257,] FALSE FALSE
#>  [258,]  TRUE FALSE
#>  [259,] FALSE FALSE
#>  [260,] FALSE FALSE
#>  [261,] FALSE FALSE
#>  [262,] FALSE FALSE
#>  [263,] FALSE  TRUE
#>  [264,] FALSE FALSE
#>  [265,] FALSE FALSE
#>  [266,] FALSE FALSE
#>  [267,] FALSE FALSE
#>  [268,] FALSE FALSE
#>  [269,] FALSE FALSE
#>  [270,] FALSE FALSE
#>  [271,] FALSE FALSE
#>  [272,] FALSE FALSE
#>  [273,]  TRUE FALSE
#>  [274,] FALSE FALSE
#>  [275,] FALSE FALSE
#>  [276,] FALSE FALSE
#>  [277,] FALSE FALSE
#>  [278,] FALSE FALSE
#>  [279,] FALSE FALSE
#>  [280,] FALSE  TRUE
#>  [281,] FALSE FALSE
#>  [282,] FALSE FALSE
#>  [283,] FALSE FALSE
#>  [284,]  TRUE FALSE
#>  [285,] FALSE FALSE
#>  [286,] FALSE FALSE
#>  [287,] FALSE FALSE
#>  [288,] FALSE FALSE
#>  [289,] FALSE FALSE
#>  [290,] FALSE  TRUE
#>  [291,] FALSE FALSE
#>  [292,] FALSE FALSE
#>  [293,] FALSE FALSE
#>  [294,]  TRUE FALSE
#>  [295,] FALSE  TRUE
#>  [296,] FALSE FALSE
#>  [297,] FALSE FALSE
#>  [298,] FALSE FALSE
#>  [299,] FALSE FALSE
#>  [300,] FALSE FALSE
#>  [301,] FALSE FALSE
#>  [302,] FALSE FALSE
#>  [303,] FALSE FALSE
#>  [304,] FALSE FALSE
#>  [305,]  TRUE FALSE
#>  [306,]  TRUE FALSE
#>  [307,] FALSE FALSE
#>  [308,] FALSE FALSE
#>  [309,] FALSE FALSE
#>  [310,] FALSE FALSE
#>  [311,] FALSE FALSE
#>  [312,] FALSE FALSE
#>  [313,] FALSE FALSE
#>  [314,] FALSE FALSE
#>  [315,] FALSE FALSE
#>  [316,] FALSE FALSE
#>  [317,] FALSE FALSE
#>  [318,] FALSE FALSE
#>  [319,] FALSE FALSE
#>  [320,] FALSE FALSE
#>  [321,] FALSE FALSE
#>  [322,] FALSE FALSE
#>  [323,] FALSE FALSE
#>  [324,] FALSE  TRUE
#>  [325,] FALSE FALSE
#>  [326,] FALSE FALSE
#>  [327,] FALSE FALSE
#>  [328,] FALSE FALSE
#>  [329,] FALSE FALSE
#>  [330,] FALSE FALSE
#>  [331,] FALSE FALSE
#>  [332,] FALSE FALSE
#>  [333,] FALSE FALSE
#>  [334,]  TRUE FALSE
#>  [335,] FALSE FALSE
#>  [336,] FALSE FALSE
#>  [337,] FALSE FALSE
#>  [338,] FALSE FALSE
#>  [339,] FALSE FALSE
#>  [340,] FALSE FALSE
#>  [341,] FALSE FALSE
#>  [342,] FALSE FALSE
#>  [343,] FALSE FALSE
#>  [344,] FALSE FALSE
#>  [345,] FALSE FALSE
#>  [346,] FALSE FALSE
#>  [347,] FALSE FALSE
#>  [348,] FALSE  TRUE
#>  [349,] FALSE FALSE
#>  [350,]  TRUE FALSE
#>  [351,] FALSE FALSE
#>  [352,] FALSE FALSE
#>  [353,] FALSE FALSE
#>  [354,]  TRUE FALSE
#>  [355,] FALSE FALSE
#>  [356,] FALSE FALSE
#>  [357,] FALSE FALSE
#>  [358,] FALSE FALSE
#>  [359,] FALSE FALSE
#>  [360,] FALSE FALSE
#>  [361,] FALSE FALSE
#>  [362,] FALSE FALSE
#>  [363,] FALSE FALSE
#>  [364,] FALSE  TRUE
#>  [365,] FALSE FALSE
#>  [366,] FALSE FALSE
#>  [367,] FALSE FALSE
#>  [368,] FALSE FALSE
#>  [369,] FALSE FALSE
#>  [370,] FALSE FALSE
#>  [371,] FALSE FALSE
#>  [372,] FALSE FALSE
#>  [373,] FALSE FALSE
#>  [374,] FALSE FALSE
#>  [375,] FALSE FALSE
#>  [376,] FALSE FALSE
#>  [377,] FALSE FALSE
#>  [378,] FALSE FALSE
#>  [379,] FALSE FALSE
#>  [380,] FALSE FALSE
#>  [381,] FALSE FALSE
#>  [382,] FALSE FALSE
#>  [383,] FALSE FALSE
#>  [384,] FALSE FALSE
#>  [385,] FALSE  TRUE
#>  [386,] FALSE FALSE
#>  [387,] FALSE FALSE
#>  [388,] FALSE FALSE
#>  [389,] FALSE FALSE
#>  [390,] FALSE FALSE
#>  [391,]  TRUE FALSE
#>  [392,] FALSE FALSE
#>  [393,] FALSE FALSE
#>  [394,]  TRUE FALSE
#>  [395,] FALSE FALSE
#>  [396,]  TRUE FALSE
#>  [397,] FALSE FALSE
#>  [398,] FALSE FALSE
#>  [399,] FALSE  TRUE
#>  [400,] FALSE FALSE
#>  [401,] FALSE  TRUE
#>  [402,] FALSE FALSE
#>  [403,] FALSE  TRUE
#>  [404,] FALSE FALSE
#>  [405,] FALSE FALSE
#>  [406,] FALSE FALSE
#>  [407,] FALSE FALSE
#>  [408,] FALSE FALSE
#>  [409,] FALSE FALSE
#>  [410,] FALSE FALSE
#>  [411,] FALSE FALSE
#>  [412,]  TRUE FALSE
#>  [413,] FALSE FALSE
#>  [414,] FALSE FALSE
#>  [415,] FALSE FALSE
#>  [416,] FALSE FALSE
#>  [417,] FALSE FALSE
#>  [418,] FALSE FALSE
#>  [419,] FALSE FALSE
#>  [420,] FALSE FALSE
#>  [421,] FALSE FALSE
#>  [422,] FALSE FALSE
#>  [423,] FALSE FALSE
#>  [424,] FALSE  TRUE
#>  [425,] FALSE FALSE
#>  [426,] FALSE FALSE
#>  [427,] FALSE FALSE
#>  [428,] FALSE FALSE
#>  [429,] FALSE FALSE
#>  [430,] FALSE FALSE
#>  [431,] FALSE FALSE
#>  [432,] FALSE FALSE
#>  [433,] FALSE FALSE
#>  [434,] FALSE  TRUE
#>  [435,] FALSE FALSE
#>  [436,] FALSE FALSE
#>  [437,] FALSE FALSE
#>  [438,] FALSE FALSE
#>  [439,] FALSE FALSE
#>  [440,] FALSE FALSE
#>  [441,] FALSE FALSE
#>  [442,] FALSE FALSE
#>  [443,] FALSE FALSE
#>  [444,] FALSE  TRUE
#>  [445,] FALSE FALSE
#>  [446,]  TRUE  TRUE
#>  [447,] FALSE FALSE
#>  [448,] FALSE  TRUE
#>  [449,] FALSE FALSE
#>  [450,] FALSE  TRUE
#>  [451,] FALSE  TRUE
#>  [452,] FALSE FALSE
#>  [453,] FALSE FALSE
#>  [454,] FALSE FALSE
#>  [455,] FALSE FALSE
#>  [456,] FALSE FALSE
#>  [457,] FALSE FALSE
#>  [458,] FALSE FALSE
#>  [459,] FALSE FALSE
#>  [460,] FALSE FALSE
#>  [461,] FALSE FALSE
#>  [462,] FALSE FALSE
#>  [463,] FALSE  TRUE
#>  [464,] FALSE FALSE
#>  [465,] FALSE FALSE
#>  [466,] FALSE FALSE
#>  [467,] FALSE FALSE
#>  [468,] FALSE FALSE
#>  [469,] FALSE FALSE
#>  [470,] FALSE FALSE
#>  [471,] FALSE FALSE
#>  [472,] FALSE FALSE
#>  [473,]  TRUE FALSE
#>  [474,]  TRUE FALSE
#>  [475,] FALSE FALSE
#>  [476,]  TRUE FALSE
#>  [477,] FALSE FALSE
#>  [478,] FALSE FALSE
#>  [479,] FALSE  TRUE
#>  [480,] FALSE FALSE
#>  [481,] FALSE FALSE
#>  [482,] FALSE FALSE
#>  [483,] FALSE FALSE
#>  [484,] FALSE FALSE
#>  [485,] FALSE FALSE
#>  [486,] FALSE  TRUE
#>  [487,] FALSE FALSE
#>  [488,] FALSE FALSE
#>  [489,] FALSE FALSE
#>  [490,] FALSE FALSE
#>  [491,]  TRUE FALSE
#>  [492,] FALSE FALSE
#>  [493,]  TRUE FALSE
#>  [494,] FALSE FALSE
#>  [495,] FALSE FALSE
#>  [496,] FALSE  TRUE
#>  [497,] FALSE  TRUE
#>  [498,] FALSE FALSE
#>  [499,] FALSE FALSE
#>  [500,]  TRUE FALSE
#>  [501,] FALSE FALSE
#>  [502,] FALSE FALSE
#>  [503,]  TRUE FALSE
#>  [504,] FALSE FALSE
#>  [505,] FALSE FALSE
#>  [506,] FALSE FALSE
#>  [507,] FALSE FALSE
#>  [508,] FALSE  TRUE
#>  [509,] FALSE FALSE
#>  [510,] FALSE FALSE
#>  [511,] FALSE FALSE
#>  [512,] FALSE FALSE
#>  [513,] FALSE FALSE
#>  [514,]  TRUE FALSE
#>  [515,]  TRUE FALSE
#>  [516,] FALSE FALSE
#>  [517,]  TRUE FALSE
#>  [518,] FALSE FALSE
#>  [519,] FALSE  TRUE
#>  [520,]  TRUE FALSE
#>  [521,] FALSE FALSE
#>  [522,] FALSE FALSE
#>  [523,] FALSE FALSE
#>  [524,] FALSE  TRUE
#>  [525,] FALSE FALSE
#>  [526,] FALSE FALSE
#>  [527,]  TRUE FALSE
#>  [528,] FALSE FALSE
#>  [529,] FALSE FALSE
#>  [530,] FALSE FALSE
#>  [531,] FALSE FALSE
#>  [532,] FALSE FALSE
#>  [533,] FALSE FALSE
#>  [534,] FALSE  TRUE
#>  [535,] FALSE FALSE
#>  [536,] FALSE  TRUE
#>  [537,]  TRUE  TRUE
#>  [538,] FALSE FALSE
#>  [539,] FALSE FALSE
#>  [540,] FALSE FALSE
#>  [541,]  TRUE FALSE
#>  [542,] FALSE FALSE
#>  [543,] FALSE FALSE
#>  [544,] FALSE FALSE
#>  [545,]  TRUE FALSE
#>  [546,] FALSE FALSE
#>  [547,] FALSE FALSE
#>  [548,] FALSE FALSE
#>  [549,] FALSE FALSE
#>  [550,] FALSE FALSE
#>  [551,] FALSE FALSE
#>  [552,]  TRUE FALSE
#>  [553,] FALSE FALSE
#>  [554,]  TRUE FALSE
#>  [555,]  TRUE FALSE
#>  [556,]  TRUE FALSE
#>  [557,] FALSE FALSE
#>  [558,] FALSE FALSE
#>  [559,] FALSE FALSE
#>  [560,] FALSE FALSE
#>  [561,] FALSE  TRUE
#>  [562,]  TRUE FALSE
#>  [563,] FALSE FALSE
#>  [564,] FALSE FALSE
#>  [565,] FALSE FALSE
#>  [566,] FALSE FALSE
#>  [567,] FALSE FALSE
#>  [568,] FALSE FALSE
#>  [569,] FALSE FALSE
#>  [570,] FALSE FALSE
#>  [571,] FALSE FALSE
#>  [572,] FALSE  TRUE
#>  [573,] FALSE  TRUE
#>  [574,] FALSE FALSE
#>  [575,] FALSE FALSE
#>  [576,] FALSE FALSE
#>  [577,] FALSE FALSE
#>  [578,] FALSE FALSE
#>  [579,] FALSE FALSE
#>  [580,]  TRUE FALSE
#>  [581,] FALSE FALSE
#>  [582,] FALSE FALSE
#>  [583,] FALSE FALSE
#>  [584,] FALSE FALSE
#>  [585,] FALSE FALSE
#>  [586,] FALSE FALSE
#>  [587,] FALSE FALSE
#>  [588,] FALSE FALSE
#>  [589,]  TRUE FALSE
#>  [590,] FALSE FALSE
#>  [591,] FALSE FALSE
#>  [592,]  TRUE FALSE
#>  [593,] FALSE FALSE
#>  [594,] FALSE FALSE
#>  [595,] FALSE FALSE
#>  [596,] FALSE FALSE
#>  [597,] FALSE FALSE
#>  [598,] FALSE FALSE
#>  [599,] FALSE  TRUE
#>  [600,] FALSE FALSE
#>  [601,]  TRUE FALSE
#>  [602,]  TRUE FALSE
#>  [603,] FALSE FALSE
#>  [604,] FALSE FALSE
#>  [605,]  TRUE FALSE
#>  [606,] FALSE FALSE
#>  [607,] FALSE FALSE
#>  [608,] FALSE FALSE
#>  [609,] FALSE FALSE
#>  [610,] FALSE  TRUE
#>  [611,] FALSE FALSE
#>  [612,] FALSE FALSE
#>  [613,] FALSE FALSE
#>  [614,] FALSE FALSE
#>  [615,] FALSE FALSE
#>  [616,] FALSE FALSE
#>  [617,] FALSE FALSE
#>  [618,] FALSE FALSE
#>  [619,] FALSE FALSE
#>  [620,] FALSE FALSE
#>  [621,] FALSE FALSE
#>  [622,] FALSE FALSE
#>  [623,] FALSE FALSE
#>  [624,] FALSE FALSE
#>  [625,] FALSE FALSE
#>  [626,] FALSE FALSE
#>  [627,] FALSE FALSE
#>  [628,] FALSE FALSE
#>  [629,] FALSE FALSE
#>  [630,] FALSE FALSE
#>  [631,] FALSE FALSE
#>  [632,] FALSE FALSE
#>  [633,] FALSE FALSE
#>  [634,] FALSE FALSE
#>  [635,] FALSE FALSE
#>  [636,] FALSE FALSE
#>  [637,] FALSE FALSE
#>  [638,]  TRUE  TRUE
#>  [639,] FALSE FALSE
#>  [640,] FALSE FALSE
#>  [641,] FALSE  TRUE
#>  [642,] FALSE FALSE
#>  [643,] FALSE FALSE
#>  [644,]  TRUE FALSE
#>  [645,] FALSE  TRUE
#>  [646,] FALSE FALSE
#>  [647,] FALSE  TRUE
#>  [648,]  TRUE FALSE
#>  [649,] FALSE FALSE
#>  [650,] FALSE FALSE
#>  [651,] FALSE FALSE
#>  [652,] FALSE  TRUE
#>  [653,] FALSE FALSE
#>  [654,]  TRUE FALSE
#>  [655,] FALSE FALSE
#>  [656,] FALSE FALSE
#>  [657,] FALSE  TRUE
#>  [658,]  TRUE  TRUE
#>  [659,] FALSE  TRUE
#>  [660,]  TRUE  TRUE
#>  [661,] FALSE FALSE
#>  [662,] FALSE FALSE
#>  [663,] FALSE FALSE
#>  [664,] FALSE  TRUE
#>  [665,] FALSE  TRUE
#>  [666,] FALSE FALSE
#>  [667,] FALSE FALSE
#>  [668,] FALSE  TRUE
#>  [669,] FALSE FALSE
#>  [670,] FALSE FALSE
#>  [671,] FALSE  TRUE
#>  [672,] FALSE FALSE
#>  [673,]  TRUE FALSE
#>  [674,] FALSE FALSE
#>  [675,] FALSE FALSE
#>  [676,] FALSE FALSE
#>  [677,] FALSE FALSE
#>  [678,] FALSE FALSE
#>  [679,] FALSE FALSE
#>  [680,] FALSE FALSE
#>  [681,]  TRUE FALSE
#>  [682,] FALSE FALSE
#>  [683,] FALSE FALSE
#>  [684,] FALSE FALSE
#>  [685,] FALSE FALSE
#>  [686,] FALSE FALSE
#>  [687,] FALSE FALSE
#>  [688,] FALSE FALSE
#>  [689,] FALSE FALSE
#>  [690,] FALSE FALSE
#>  [691,] FALSE FALSE
#>  [692,] FALSE FALSE
#>  [693,] FALSE FALSE
#>  [694,] FALSE FALSE
#>  [695,] FALSE FALSE
#>  [696,] FALSE FALSE
#>  [697,] FALSE  TRUE
#>  [698,] FALSE FALSE
#>  [699,] FALSE FALSE
#>  [700,] FALSE FALSE
#>  [701,] FALSE FALSE
#>  [702,] FALSE FALSE
#>  [703,] FALSE FALSE
#>  [704,]  TRUE FALSE
#>  [705,] FALSE FALSE
#>  [706,]  TRUE FALSE
#>  [707,] FALSE FALSE
#>  [708,] FALSE FALSE
#>  [709,] FALSE FALSE
#>  [710,] FALSE FALSE
#>  [711,]  TRUE FALSE
#>  [712,] FALSE  TRUE
#>  [713,] FALSE FALSE
#>  [714,] FALSE FALSE
#>  [715,] FALSE FALSE
#>  [716,] FALSE FALSE
#>  [717,] FALSE FALSE
#>  [718,] FALSE FALSE
#>  [719,] FALSE FALSE
#>  [720,] FALSE FALSE
#>  [721,] FALSE FALSE
#>  [722,] FALSE FALSE
#>  [723,] FALSE FALSE
#>  [724,]  TRUE FALSE
#>  [725,] FALSE FALSE
#>  [726,] FALSE  TRUE
#>  [727,]  TRUE FALSE
#>  [728,] FALSE FALSE
#>  [729,] FALSE FALSE
#>  [730,] FALSE FALSE
#>  [731,]  TRUE FALSE
#>  [732,] FALSE FALSE
#>  [733,]  TRUE FALSE
#>  [734,] FALSE FALSE
#>  [735,] FALSE FALSE
#>  [736,] FALSE FALSE
#>  [737,]  TRUE FALSE
#>  [738,] FALSE FALSE
#>  [739,] FALSE FALSE
#>  [740,] FALSE  TRUE
#>  [741,] FALSE FALSE
#>  [742,]  TRUE FALSE
#>  [743,] FALSE FALSE
#>  [744,] FALSE FALSE
#>  [745,] FALSE FALSE
#>  [746,] FALSE FALSE
#>  [747,] FALSE FALSE
#>  [748,] FALSE FALSE
#>  [749,] FALSE  TRUE
#>  [750,] FALSE FALSE
#>  [751,]  TRUE FALSE
#>  [752,] FALSE FALSE
#>  [753,] FALSE FALSE
#>  [754,] FALSE FALSE
#>  [755,]  TRUE FALSE
#>  [756,]  TRUE FALSE
#>  [757,] FALSE FALSE
#>  [758,] FALSE FALSE
#>  [759,] FALSE FALSE
#>  [760,]  TRUE FALSE
#>  [761,] FALSE FALSE
#>  [762,] FALSE FALSE
#>  [763,] FALSE FALSE
#>  [764,] FALSE FALSE
#>  [765,] FALSE FALSE
#>  [766,] FALSE FALSE
#>  [767,] FALSE FALSE
#>  [768,] FALSE FALSE
#>  [769,] FALSE FALSE
#>  [770,] FALSE FALSE
#>  [771,]  TRUE FALSE
#>  [772,] FALSE FALSE
#>  [773,] FALSE FALSE
#>  [774,] FALSE FALSE
#>  [775,] FALSE FALSE
#>  [776,] FALSE FALSE
#>  [777,] FALSE  TRUE
#>  [778,] FALSE  TRUE
#>  [779,] FALSE FALSE
#>  [780,] FALSE FALSE
#>  [781,] FALSE FALSE
#>  [782,] FALSE FALSE
#>  [783,] FALSE FALSE
#>  [784,] FALSE FALSE
#>  [785,] FALSE FALSE
#>  [786,] FALSE FALSE
#>  [787,]  TRUE FALSE
#>  [788,] FALSE FALSE
#>  [789,] FALSE FALSE
#>  [790,] FALSE FALSE
#>  [791,] FALSE FALSE
#>  [792,] FALSE FALSE
#>  [793,] FALSE FALSE
#>  [794,] FALSE FALSE
#>  [795,] FALSE FALSE
#>  [796,] FALSE FALSE
#>  [797,]  TRUE FALSE
#>  [798,] FALSE FALSE
#>  [799,] FALSE  TRUE
#>  [800,] FALSE FALSE
#>  [801,] FALSE FALSE
#>  [802,] FALSE FALSE
#>  [803,] FALSE FALSE
#>  [804,] FALSE FALSE
#>  [805,] FALSE FALSE
#>  [806,] FALSE FALSE
#>  [807,] FALSE FALSE
#>  [808,] FALSE FALSE
#>  [809,] FALSE FALSE
#>  [810,] FALSE FALSE
#>  [811,] FALSE FALSE
#>  [812,] FALSE FALSE
#>  [813,] FALSE FALSE
#>  [814,] FALSE FALSE
#>  [815,] FALSE FALSE
#>  [816,] FALSE FALSE
#>  [817,] FALSE FALSE
#>  [818,] FALSE  TRUE
#>  [819,] FALSE FALSE
#>  [820,] FALSE FALSE
#>  [821,] FALSE FALSE
#>  [822,] FALSE FALSE
#>  [823,] FALSE FALSE
#>  [824,] FALSE FALSE
#>  [825,]  TRUE FALSE
#>  [826,] FALSE FALSE
#>  [827,] FALSE FALSE
#>  [828,]  TRUE FALSE
#>  [829,] FALSE FALSE
#>  [830,] FALSE FALSE
#>  [831,] FALSE FALSE
#>  [832,] FALSE FALSE
#>  [833,] FALSE FALSE
#>  [834,] FALSE FALSE
#>  [835,]  TRUE FALSE
#>  [836,] FALSE FALSE
#>  [837,] FALSE FALSE
#>  [838,] FALSE FALSE
#>  [839,] FALSE FALSE
#>  [840,] FALSE FALSE
#>  [841,] FALSE FALSE
#>  [842,] FALSE FALSE
#>  [843,] FALSE FALSE
#>  [844,] FALSE FALSE
#>  [845,] FALSE FALSE
#>  [846,] FALSE FALSE
#>  [847,] FALSE FALSE
#>  [848,] FALSE FALSE
#>  [849,] FALSE FALSE
#>  [850,] FALSE FALSE
#>  [851,] FALSE  TRUE
#>  [852,] FALSE  TRUE
#>  [853,] FALSE FALSE
#>  [854,] FALSE FALSE
#>  [855,] FALSE  TRUE
#>  [856,] FALSE FALSE
#>  [857,] FALSE FALSE
#>  [858,] FALSE FALSE
#>  [859,] FALSE FALSE
#>  [860,] FALSE  TRUE
#>  [861,] FALSE FALSE
#>  [862,] FALSE FALSE
#>  [863,] FALSE FALSE
#>  [864,] FALSE FALSE
#>  [865,]  TRUE FALSE
#>  [866,] FALSE FALSE
#>  [867,] FALSE FALSE
#>  [868,] FALSE FALSE
#>  [869,] FALSE FALSE
#>  [870,]  TRUE FALSE
#>  [871,] FALSE FALSE
#>  [872,] FALSE FALSE
#>  [873,] FALSE FALSE
#>  [874,] FALSE FALSE
#>  [875,] FALSE FALSE
#>  [876,] FALSE FALSE
#>  [877,] FALSE FALSE
#>  [878,] FALSE FALSE
#>  [879,] FALSE FALSE
#>  [880,] FALSE FALSE
#>  [881,]  TRUE FALSE
#>  [882,] FALSE FALSE
#>  [883,] FALSE FALSE
#>  [884,] FALSE  TRUE
#>  [885,] FALSE FALSE
#>  [886,] FALSE FALSE
#>  [887,] FALSE FALSE
#>  [888,] FALSE  TRUE
#>  [889,] FALSE FALSE
#>  [890,] FALSE FALSE
#>  [891,] FALSE FALSE
#>  [892,] FALSE FALSE
#>  [893,] FALSE FALSE
#>  [894,] FALSE FALSE
#>  [895,] FALSE FALSE
#>  [896,] FALSE FALSE
#>  [897,] FALSE  TRUE
#>  [898,] FALSE FALSE
#>  [899,]  TRUE FALSE
#>  [900,] FALSE FALSE
#>  [901,] FALSE FALSE
#>  [902,] FALSE FALSE
#>  [903,] FALSE FALSE
#>  [904,] FALSE FALSE
#>  [905,]  TRUE FALSE
#>  [906,] FALSE FALSE
#>  [907,] FALSE FALSE
#>  [908,] FALSE FALSE
#>  [909,] FALSE FALSE
#>  [910,] FALSE FALSE
#>  [911,] FALSE FALSE
#>  [912,] FALSE FALSE
#>  [913,] FALSE FALSE
#>  [914,] FALSE FALSE
#>  [915,] FALSE FALSE
#>  [916,] FALSE FALSE
#>  [917,] FALSE FALSE
#>  [918,] FALSE FALSE
#>  [919,] FALSE FALSE
#>  [920,] FALSE FALSE
#>  [921,] FALSE FALSE
#>  [922,] FALSE FALSE
#>  [923,] FALSE FALSE
#>  [924,] FALSE FALSE
#>  [925,] FALSE FALSE
#>  [926,] FALSE FALSE
#>  [927,] FALSE FALSE
#>  [928,] FALSE FALSE
#>  [929,] FALSE FALSE
#>  [930,] FALSE FALSE
#>  [931,] FALSE FALSE
#>  [932,] FALSE FALSE
#>  [933,] FALSE FALSE
#>  [934,] FALSE FALSE
#>  [935,] FALSE FALSE
#>  [936,] FALSE FALSE
#>  [937,] FALSE FALSE
#>  [938,] FALSE FALSE
#>  [939,] FALSE FALSE
#>  [940,] FALSE FALSE
#>  [941,] FALSE FALSE
#>  [942,] FALSE FALSE
#>  [943,] FALSE FALSE
#>  [944,] FALSE FALSE
#>  [945,] FALSE FALSE
#>  [946,]  TRUE FALSE
#>  [947,] FALSE FALSE
#>  [948,] FALSE FALSE
#>  [949,] FALSE FALSE
#>  [950,] FALSE FALSE
#>  [951,] FALSE FALSE
#>  [952,] FALSE FALSE
#>  [953,] FALSE  TRUE
#>  [954,] FALSE  TRUE
#>  [955,]  TRUE FALSE
#>  [956,] FALSE FALSE
#>  [957,] FALSE  TRUE
#>  [958,] FALSE FALSE
#>  [959,] FALSE FALSE
#>  [960,] FALSE FALSE
#>  [961,]  TRUE FALSE
#>  [962,] FALSE FALSE
#>  [963,] FALSE FALSE
#>  [964,] FALSE FALSE
#>  [965,] FALSE FALSE
#>  [966,] FALSE FALSE
#>  [967,]  TRUE FALSE
#>  [968,] FALSE FALSE
#>  [969,]  TRUE FALSE
#>  [970,]  TRUE FALSE
#>  [971,] FALSE FALSE
#>  [972,]  TRUE FALSE
#>  [973,] FALSE  TRUE
#>  [974,] FALSE FALSE
#>  [975,] FALSE  TRUE
#>  [976,] FALSE FALSE
#>  [977,] FALSE FALSE
#>  [978,] FALSE FALSE
#>  [979,] FALSE FALSE
#>  [980,] FALSE FALSE
#>  [981,] FALSE FALSE
#>  [982,] FALSE FALSE
#>  [983,] FALSE FALSE
#>  [984,] FALSE FALSE
#>  [985,] FALSE FALSE
#>  [986,] FALSE FALSE
#>  [987,] FALSE FALSE
#>  [988,] FALSE FALSE
#>  [989,] FALSE FALSE
#>  [990,] FALSE FALSE
#>  [991,] FALSE FALSE
#>  [992,] FALSE FALSE
#>  [993,] FALSE FALSE
#>  [994,] FALSE FALSE
#>  [995,] FALSE FALSE
#>  [996,]  TRUE FALSE
#>  [997,] FALSE FALSE
#>  [998,] FALSE FALSE
#>  [999,] FALSE FALSE
#> [1000,] FALSE FALSE
#> 
#> $no.uniques
#> [1] 425
#> 
#> $no.replications
#> [1] 108  88
#> 
#> $per.replications
#> [1] 10.8  8.8
#>