Synthetic data object summaries

Produces summaries of the synthesised variables. When more than one synthetic data set has been generated (object$m > 1), by default summaries are calculated by averaging summary values for all synthetic data copies (see msel argument).

Usage

# S3 method for synds
summary(object, msel = NULL, maxsum = 7, 
  digits = max(3, getOption("digits")-3), ...)

# S3 method for summary.synds
print(x, ...)

Arguments

object

an object of class synds; a result of a call to syn.

msel

index or indices of synthetic data copies for which a summary is desired. If NULL (default) summaries are calculated by averaging summary values for all synthetic data copies.

maxsum

integer, indicating how many levels should be shown for factors.

digits

integer, used for number formatting with format.

...

additional arguments passed to summary.

x

an object of class summary.synds.

Details

See summary for more details.

Value

An object of class summary.synds, which is a list with the following components:

m

the number of synthetic versions of the original (observed) data.

msel

index or indices of synthetic data copies for which a summary is produced. If NULL summaries are calculated by averaging summary values for all synthetic data copies.

method

a vector of synthesising methods applied to each variable in the saved synthesised data.

result

a table or a list of tabels (if more than one synthetic data set is selected) with summaries of synthesised variables.

References

Nowok, B., Raab, G.M and Dibben, C. (2016). synthpop: Bespoke creation of synthetic data in R. Journal of Statistical Software, 74(11), 1-26. doi:10.18637/jss.v074.i11 .

See also

Examples

s1 <- syn(SD2011[,c("sex","age","edu","marital")], m = 3)
#> 
#> Synthesis number 1
#> --------------------
#>  sex age edu marital
#> 
#> Synthesis number 2
#> --------------------
#>  sex age edu marital
#> 
#> Synthesis number 3
#> --------------------
#>  sex age edu marital
summary(s1)
#> Synthetic object with 3 syntheses using methods:
#>      sex      age      edu  marital 
#> "sample"   "cart"   "cart"   "cart" 
#> 
#> Summary (average) for all synthetic data sets:
#>      sex            age                              edu          
#>  MALE  :2189   Min.   :16.00   PRIMARY/NO EDUCATION    : 952.667  
#>  FEMALE:2811   1st Qu.:31.67   VOCATIONAL/GRAMMAR      :1639.000  
#>                Median :48.33   SECONDARY               :1476.000  
#>                Mean   :47.40   POST-SECONDARY OR HIGHER: 926.667  
#>                3rd Qu.:61.00   NA's                    :   5.667  
#>                Max.   :96.67                                      
#>                                                                   
#>                marital        
#>  SINGLE            :1280.667  
#>  MARRIED           :2946.333  
#>  WIDOWED           : 527.333  
#>  DIVORCED          : 204.667  
#>  LEGALLY SEPARATED :   6.667  
#>  DE FACTO SEPARATED:  22.667  
#>  NA's              :  11.667  
summary(s1, msel = c(1,3))
#> Synthetic object with 3 syntheses using methods:
#>      sex      age      edu  marital 
#> "sample"   "cart"   "cart"   "cart" 
#> 
#> Summary for synthetic data set 1:
#>      sex            age                              edu      
#>  MALE  :2168   Min.   :16.00   PRIMARY/NO EDUCATION    : 949  
#>  FEMALE:2832   1st Qu.:32.00   VOCATIONAL/GRAMMAR      :1646  
#>                Median :48.00   SECONDARY               :1450  
#>                Mean   :47.61   POST-SECONDARY OR HIGHER: 947  
#>                3rd Qu.:61.00   NA's                    :   8  
#>                Max.   :97.00                                  
#>                                                               
#>                marital    
#>  SINGLE            :1192  
#>  MARRIED           :3048  
#>  WIDOWED           : 511  
#>  DIVORCED          : 213  
#>  LEGALLY SEPARATED :   6  
#>  DE FACTO SEPARATED:  17  
#>  NA's              :  13  
#> 
#> Summary for synthetic data set 3:
#>      sex            age                              edu      
#>  MALE  :2212   Min.   :16.00   PRIMARY/NO EDUCATION    : 957  
#>  FEMALE:2788   1st Qu.:31.00   VOCATIONAL/GRAMMAR      :1646  
#>                Median :48.00   SECONDARY               :1489  
#>                Mean   :47.02   POST-SECONDARY OR HIGHER: 905  
#>                3rd Qu.:61.00   NA's                    :   3  
#>                Max.   :97.00                                  
#>                                                               
#>                marital    
#>  SINGLE            :1369  
#>  MARRIED           :2866  
#>  WIDOWED           : 512  
#>  DIVORCED          : 211  
#>  LEGALLY SEPARATED :   6  
#>  DE FACTO SEPARATED:  26  
#>  NA's              :  10