Inference from synthetic data

Combines the results of models fitted to each of the m synthetic data sets.

Usage

# S3 method for fit.synds
summary(object, population.inference = FALSE, msel = NULL,
  real.varcov = NULL, incomplete = NULL, ...)

# S3 method for summary.fit.synds
print(x, ...)

Arguments

object

an object of class fit.synds created by fitting a model to synthesised data set using function glm.synds, lm.synds,multinom.synds or polr.synds.

population.inference

a logical value indicating whether inference should be made to population quantities. If FALSE inference is made to the results that would be expected from an analysis of the original data. This option should be selected if the synthetic data are being used for exploratory analysis, but the final published results will be obtained by running code on the original confidential data. If population.inference = TRUE results would allow population inference to be made from the synthetic data. In both cases the inference will depend on the synthesising model being correct, but this can be checked by running the same analysis on the real data, see compare.fit.synds.

msel

index or indices of the synthetic datasets (1, ..., m), for which summaries of fitted models are to be produced. If NULL (default) only the summary of combined estimates is produced.

real.varcov

the estimated variance-covariance matrix of the fit of the model to the original data. This parameter is used in the function compare.fit.synds which has the original data as one of its parameters.

incomplete

Logical variable as to whether population inference for incomplete synthesis is to be used. If this is left at a NULL value it will be determined by whether the dependent variable has been synthesised. See also below as output.

...

additional parameters.

x

an object of class summary.fit.synds.

Details

The mean of the estimates from each of the m synthetic data sets yields asymptotically unbiased estimates of the coefficients if the observed data conform to the distribution used for synthesis. The standard errors are estimated differently depending whether inference is made for the results that we would expect to obtain from the observed data or for the parameters of the population that we assume the observed data are sampled from. The standard errors also differ according to whether synthetic data were produced using simple or proper synthesis (for details see Raab et al. (2017)).

Value

An object of class summary.fit.synds which is a list with the following components:

call

the original call to glm.synds or lm.synds.

proper

a logical value indicating whether synthetic data were generated using proper synthesis.

population.inference

a logical value indicating whether inference is made to population coefficients or to the results that would be expected from an analysis of the original data (see above).

incomplete

a logical value indicating whether the dependent variable in the model was not synthesised. It is derived in the synthpop implementation of the fitting functions (lm.synds, glm.synds, multinom.synds and polr.synds) and saved with the fitted object. When TRUE inference with population.inference = TRUE uses the method proposed by Reiter (2003) for what he terms partially synthetic data. This method requires multiple syntheses (m > 1). If m = 1, incomplete = TRUE and population.inference = TRUE the results will be still calculated and returned with warning. This will usually give standard errors that are larger than they should be. This method can be forced by setting incomplete = TRUE as a parameter because it can also be used for complete synthesis.

fitting.function

function used to fit the model.

m

the number of synthetic versions of the original (observed) data.

coefficients

a matrix with combined estimates. If inference is required to the results that would be obtained from an analysis of the original data, (population.inference = FALSE) the coefficients are given by xpct(Beta), the standard errors by xpct(se.Beta) and the corresponding Z-statistic by xpct(Z). If the synthetic data are to be used to make inferences to population quantities (population.inference = TRUE), the coefficients are given by Beta.syn, their standard errors by se.Beta.syn and the Z-statistic by Z.syn (see vignette on inference for more details).

n

a number of cases in the original data.

k

the number of cases in the synthesised data. Note that if k and n are not equal and population.inference = FALSE (the default), then the standard errors produced will estimate what would be expected by an analysis of the original data set of size n.

analyses

summary.glm or summary.lm object respectively or a list of m such objects.

msel

index or indices of synthetic data copies for which summaries of fitted models are produced. If NULL only a summary of combined estimates is produced.

References

Nowok, B., Raab, G.M and Dibben, C. (2016). synthpop: Bespoke creation of synthetic data in R. Journal of Statistical Software, 74(11), 1-26. doi:10.18637/jss.v074.i11 .

Raab, G.M., Nowok, B. and Dibben, C. (2017). Practical data synthesis for large samples. Journal of Privacy and Confidentiality, 7(3), 67-97. Available at: https://journalprivacyconfidentiality.org/index.php/jpc/article/view/407

Reiter, J.P. (2003) Inference for partially synthetic, public use microdata sets. Survey Methodology, 29, 181-188.

See also

Examples

ods <- SD2011[1:1000,c("sex","age","edu","ls","smoke")]
  
### simple synthesis
s1 <- syn(ods, m = 5)
#> 
#> Synthesis number 1
#> --------------------
#>  sex age edu ls smoke
#> 
#> Synthesis number 2
#> --------------------
#>  sex age edu ls smoke
#> 
#> Synthesis number 3
#> --------------------
#>  sex age edu ls smoke
#> 
#> Synthesis number 4
#> --------------------
#>  sex age edu ls smoke
#> 
#> Synthesis number 5
#> --------------------
#>  sex age edu ls smoke
f1 <- glm.synds(smoke ~ sex + age + edu + ls, data = s1, family = "binomial")
summary(f1)
#> Fit to synthetic data set with 5 syntheses. Inference to coefficients
#> and standard errors that would be obtained from the original data.
#> 
#> Call:
#> glm.synds(formula = smoke ~ sex + age + edu + ls, family = "binomial", 
#>     data = s1)
#> 
#> Combined estimates:
#>                              xpct(Beta) xpct(se.Beta) xpct(z) Pr(>|xpct(z)|)   
#> (Intercept)                   1.0972224     0.6029214  1.8198       0.068783 . 
#> sexFEMALE                     0.2277368     0.1548988  1.4702       0.141500   
#> age                           0.0127309     0.0045113  2.8220       0.004773 **
#> eduVOCATIONAL/GRAMMAR        -0.4976659     0.2470774 -2.0142       0.043987 * 
#> eduSECONDARY                  0.0958399     0.2627419  0.3648       0.715284   
#> eduPOST-SECONDARY OR HIGHER   0.4088187     0.3086988  1.3243       0.185394   
#> lsPLEASED                    -0.4063205     0.5204304 -0.7807       0.434956   
#> lsMOSTLY SATISFIED           -0.6192484     0.5225364 -1.1851       0.235985   
#> lsMIXED                      -0.8023550     0.5408508 -1.4835       0.137940   
#> lsMOSTLY DISSATISFIED        -1.1562755     0.5812813 -1.9892       0.046681 * 
#> lsUNHAPPY                     1.9696772   229.0313194  0.0086       0.993138   
#> lsTERRIBLE                   -1.6394358     0.8588367 -1.9089       0.056275 . 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
summary(f1, population.inference = TRUE)
#> Fit to synthetic data set with 5 syntheses. Inference to population
#> coefficients when all variables in the model are synthesised.
#> 
#> Call:
#> glm.synds(formula = smoke ~ sex + age + edu + ls, family = "binomial", 
#>     data = s1)
#> 
#> Combined estimates:
#>                                Beta.syn se.Beta.syn   z.syn Pr(>|z.syn|)   
#> (Intercept)                   1.0972224   0.6604673  1.6613     0.096657 . 
#> sexFEMALE                     0.2277368   0.1696832  1.3421     0.179554   
#> age                           0.0127309   0.0049419  2.5761     0.009992 **
#> eduVOCATIONAL/GRAMMAR        -0.4976659   0.2706598 -1.8387     0.065957 . 
#> eduSECONDARY                  0.0958399   0.2878193  0.3330     0.739145   
#> eduPOST-SECONDARY OR HIGHER   0.4088187   0.3381626  1.2089     0.226685   
#> lsPLEASED                    -0.4063205   0.5701030 -0.7127     0.476023   
#> lsMOSTLY SATISFIED           -0.6192484   0.5724099 -1.0818     0.279330   
#> lsMIXED                      -0.8023550   0.5924723 -1.3542     0.175657   
#> lsMOSTLY DISSATISFIED        -1.1562755   0.6367618 -1.8159     0.069391 . 
#> lsUNHAPPY                     1.9696772 250.8912401  0.0079     0.993736   
#> lsTERRIBLE                   -1.6394358   0.9408084 -1.7426     0.081407 . 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
  
### proper synthesis
s2 <- syn(ods, m = 5, method = "parametric", proper = TRUE)
#> 
#> Synthesis number 1
#> --------------------
#>  sex age edu ls
#> Warning: NaNs produced
#>  smoke
#> 
#> Synthesis number 2
#> --------------------
#>  sex age edu ls
#> Warning: NaNs produced
#>  smoke
#> 
#> Synthesis number 3
#> --------------------
#>  sex age edu ls smoke
#> 
#> Synthesis number 4
#> --------------------
#>  sex age edu ls smoke
#> 
#> Synthesis number 5
#> --------------------
#>  sex age edu ls smoke
f2 <- glm.synds(smoke ~ sex + age + edu + ls, data = s2, family = "binomial")
summary(f2)
#> Fit to synthetic data set with 5 syntheses. Inference to coefficients
#> and standard errors that would be obtained from the original data.
#> 
#> Call:
#> glm.synds(formula = smoke ~ sex + age + edu + ls, family = "binomial", 
#>     data = s2)
#> 
#> Combined estimates:
#>                              xpct(Beta) xpct(se.Beta) xpct(z) Pr(>|xpct(z)|)   
#> (Intercept)                   0.8511737     0.5905332  1.4414       0.149482   
#> sexFEMALE                     0.4127722     0.1573828  2.6227       0.008723 **
#> age                           0.0122694     0.0045998  2.6674       0.007645 **
#> eduVOCATIONAL/GRAMMAR        -0.2770027     0.2443218 -1.1338       0.256895   
#> eduSECONDARY                  0.2164908     0.2595280  0.8342       0.404184   
#> eduPOST-SECONDARY OR HIGHER   0.1803386     0.2961063  0.6090       0.542502   
#> lsPLEASED                    -0.2520522     0.4971541 -0.5070       0.612162   
#> lsMOSTLY SATISFIED           -0.2960799     0.5002483 -0.5919       0.553940   
#> lsMIXED                      -0.9503364     0.5122799 -1.8551       0.063580 . 
#> lsMOSTLY DISSATISFIED        -1.6209590     0.5544799 -2.9234       0.003462 **
#> lsUNHAPPY                     3.0501106   214.3335806  0.0142       0.988646   
#> lsTERRIBLE                   -4.6552078   195.6936965 -0.0238       0.981022   
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
summary(f2, population.inference = TRUE)
#> Fit to synthetic data set with 5 syntheses. Inference to population
#> coefficients when all variables in the model are synthesised.
#> 
#> Call:
#> glm.synds(formula = smoke ~ sex + age + edu + ls, family = "binomial", 
#>     data = s2)
#> 
#> Combined estimates:
#>                                Beta.syn se.Beta.syn   z.syn Pr(>|z.syn|)  
#> (Intercept)                   0.8511737   0.6987283  1.2182      0.22316  
#> sexFEMALE                     0.4127722   0.1862178  2.2166      0.02665 *
#> age                           0.0122694   0.0054426  2.2543      0.02417 *
#> eduVOCATIONAL/GRAMMAR        -0.2770027   0.2890855 -0.9582      0.33796  
#> eduSECONDARY                  0.2164908   0.3070776  0.7050      0.48081  
#> eduPOST-SECONDARY OR HIGHER   0.1803386   0.3503577  0.5147      0.60674  
#> lsPLEASED                    -0.2520522   0.5882407 -0.4285      0.66830  
#> lsMOSTLY SATISFIED           -0.2960799   0.5919017 -0.5002      0.61692  
#> lsMIXED                      -0.9503364   0.6061378 -1.5679      0.11691  
#> lsMOSTLY DISSATISFIED        -1.6209590   0.6560695 -2.4707      0.01348 *
#> lsUNHAPPY                     3.0501106 253.6029126  0.0120      0.99040  
#> lsTERRIBLE                   -4.6552078 231.5479043 -0.0201      0.98396  
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1